requirements.txt
fastapi>=0.100.0
uvicorn>=0.24.0
python-multipart>=0.0.7
google-cloud-storage>=3.13.0
google-cloud-firestore>=2.11.0
openai>=1.9.0
python-dotenv>=1.0.0
pydantic>=2.6.0

